Total and End receipt
Wrote the code to get the Subtotal, Tax, and Total as well as formatting the receipt and doing the end the receipt. Changed it where the inputs come before the beginning of the receipt so it looks cleaner. Also made changes to input section that allow the store owner to ad items and id's without breaking the code.